requirements.txt
# Core Application
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector Database
chromadb>=0.4.0

# OpenAI API
openai>=1.0.0